+++ crates/holochain/tests/integrity_zome/mod.rs
@@ -13,7 +13,7 @@ use holochain_types::dna::ZomeDependency;
 use holochain_types::dna::ZomeLocation;
 use holochain_types::dna::ZomeManifest;
 use holochain_types::prelude::DnaWasm;
-use holochain_types::prelude::HotSwapCoordinatorsPayload;
+use holochain_types::prelude::UpdateCoordinatorsPayload;
 use holochain_wasm_test_utils::TestCoordinatorWasm;
 use holochain_wasm_test_utils::TestIntegrityWasm;
 use holochain_zome_types::CoordinatorZome;
@@ -26,7 +26,7 @@ use holochain_zome_types::ZomeDef;
 use mr_bundle::Bundle;
 
 #[tokio::test(flavor = "multi_thread")]
-async fn test_coordinator_zome_hot_swap() {
+async fn test_coordinator_zome_update() {
     let mut conductor = SweetConductor::from_config(Default::default()).await;
     let (dna, _, _) = SweetDnaFile::unique_from_zomes(
         vec![TestIntegrityWasm::IntegrityZome],
@@ -68,7 +68,7 @@ async fn test_coordinator_zome_hot_swap() {
 
     println!("Hot swap the coordinator zomes for a totally different coordinator zome (conductor is still running)");
     conductor
-        .hot_swap_coordinators(
+        .update_coordinators(
             &dna_hash,
             vec![CoordinatorZome::from(TestCoordinatorWasm::CoordinatorZomeUpdate).into_inner()],
             vec![TestCoordinatorWasm::CoordinatorZomeUpdate.into()],
@@ -91,7 +91,7 @@ async fn test_coordinator_zome_hot_swap() {
 }
 
 #[tokio::test(flavor = "multi_thread")]
-async fn test_coordinator_zome_hot_swap_multi_integrity() {
+async fn test_coordinator_zome_update_multi_integrity() {
     let mut conductor = SweetConductor::from_config(Default::default()).await;
     let mut second_integrity = IntegrityZome::from(TestIntegrityWasm::IntegrityZome);
     second_integrity.zome_name_mut().0 = "2".into();
@@ -169,7 +169,7 @@ async fn test_coordinator_zome_hot_swap_multi_integrity() {
 
     // Add a completely new coordinator with the same dependency
     conductor
-        .hot_swap_coordinators(
+        .update_coordinators(
             &dna_hash,
             vec![CoordinatorZome::from(TestCoordinatorWasm::CoordinatorZomeUpdate).into_inner()],
             vec![TestCoordinatorWasm::CoordinatorZomeUpdate.into()],
@@ -197,7 +197,7 @@ async fn test_coordinator_zome_hot_swap_multi_integrity() {
     .into();
 
     conductor
-        .hot_swap_coordinators(
+        .update_coordinators(
             &dna_hash,
             vec![("2_coord".into(), new_coordinator)],
             vec![TestCoordinatorWasm::CoordinatorZomeUpdate.into()],
@@ -213,7 +213,7 @@ async fn test_coordinator_zome_hot_swap_multi_integrity() {
 }
 
 #[tokio::test(flavor = "multi_thread")]
-async fn test_hot_swap_admin_interface() {
+async fn test_update_admin_interface() {
     let mut conductor = SweetConductor::from_config(Default::default()).await;
     let (dna, _, _) = SweetDnaFile::unique_from_zomes(
         vec![TestIntegrityWasm::IntegrityZome],
@@ -258,11 +258,11 @@ async fn test_hot_swap_admin_interface() {
     .unwrap()
     .into();
 
-    let req = HotSwapCoordinatorsPayload {
+    let req = UpdateCoordinatorsPayload {
         dna_hash,
         source: holochain_types::prelude::CoordinatorSource::Bundle(Box::new(source)),
     };
-    let req = AdminRequest::HotSwapCoordinators(Box::new(req));
+    let req = AdminRequest::UpdateCoordinators(Box::new(req));
     let r = admin_api.handle_admin_request(req).await;
-    assert!(matches!(r, AdminResponse::CoordinatorsHotSwapped));
+    assert!(matches!(r, AdminResponse::CoordinatorsUpdated));
 }